mbimcli: fix warnings with -Wsign-compare
  mbimcli-basic-connect.c: In function ‘connect_session_id_parse’:
  mbimcli-basic-connect.c:1107:54: warning: comparison of integer expressions of different signedness: ‘long int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
   1107 |     if (errno || n < 0 || n > 255 || ((endptr - str) < strlen (str))) {
        |                                                      ^

  mbimcli-ms-basic-connect-extensions.c: In function ‘session_id_parse’:
  mbimcli-ms-basic-connect-extensions.c:114:54: warning: comparison of integer expressions of different signedness: ‘long int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
    114 |     if (errno || n < 0 || n > 255 || ((endptr - str) < strlen (str))) {
        |                                                      ^

+++ src/mbimcli/mbimcli-basic-connect.c
@@ -1104,7 +1104,7 @@ connect_session_id_parse (const gchar  *str,
 
     errno = 0;
     n = g_ascii_strtoll (str, &endptr, 10);
-    if (errno || n < 0 || n > 255 || ((endptr - str) < strlen (str))) {
+    if (errno || n < 0 || n > 255 || ((size_t)(endptr - str) < strlen (str))) {
         g_set_error (error,
                      MBIM_CORE_ERROR,
                      MBIM_CORE_ERROR_FAILED,
